refactor(cli): split skills adapters and cli helpers into modules
